combine the integration tests for detailed monitoring and dashboard and alarms and rename the section as monitoring
