+++ tests/integration-tests/configs/common/common.yaml
@@ -206,8 +206,8 @@ createami:
       - regions: ["ca-central-1"]
         instances: {{ common.INSTANCES_DEFAULT_X86 }}
         oss: ["alinux2"]
-dashboard_and_alarms:
-  test_dashboard_and_alarms.py::test_dashboard_and_alarms:
+monitoring:
+  test_monitoring.py::test_monitoring:
     dimensions:
       - regions: ["ap-northeast-2"]
         instances: {{ common.INSTANCES_DEFAULT_X86 }}
@@ -218,13 +218,6 @@ dashboard_and_alarms:
       - regions: ["af-south-1"]
         instances: {{ common.INSTANCES_DEFAULT_X86 }}
         oss: ["alinux2", "ubuntu2004"]
-        schedulers: ["slurm"]
-detailed_monitoring:
-  test_detailed_monitoring.py::test_detailed_monitoring:
-    dimensions:
-      - regions: ["af-south-1"]
-        instances: {{ common.INSTANCES_DEFAULT_X86 }}
-        oss: ["alinux2"]
         schedulers: ["slurm"]
 dcv:
   test_dcv.py::test_dcv_configuration:
