+++ tests/integration-tests/configs/new_region.yaml
@@ -62,8 +62,8 @@ test-suites:
           instances: {{ common.INSTANCES_DEFAULT_X86 }}
           oss: ["alinux2"]
           schedulers: ["slurm"]
-  dashboard_and_alarms:
-    test_dashboard_and_alarms.py::test_dashboard_and_alarms:
+  monitoring:
+    test_monitoring.py::test_monitoring:
       dimensions:
         - regions: {{ NEW_REGION }}
           instances: {{ common.INSTANCES_DEFAULT_X86 }}
