--- tests/integration-tests/tests/detailed_monitoring/test_detailed_monitoring/test_detailed_monitoring/pcluster.config.yaml
@@ -1,25 +1,0 @@
-Image:
-  Os: {{ os }}
-HeadNode:
-  InstanceType: {{ instance }}
-  Networking:
-    SubnetId: {{ public_subnet_id }}
-  Ssh:
-    KeyName: {{ key_name }}
-  Imds:
-    Secured: {{ imds_secured }}
-Scheduling:
-  Scheduler: {{ scheduler }}
-  {% if scheduler == "awsbatch" %}AwsBatchQueues:{% else %}SlurmQueues:{% endif %}
-    - Name: queue1
-      Networking:
-        SubnetIds:
-          - {{ public_subnet_id }}
-      ComputeResources:
-        - Name: compute-resource1
-          Instances:
-            - InstanceType: {{ instance }}
-          # set MinCount to 1 in order to test the settings of detailed monitoring
-          MinCount: 1
-Monitoring:
-  DetailedMonitoring: {{ detailed_monitoring_enabled }}
