+++ tests/integration-tests/tests/monitoring/test_monitoring/test_monitoring/pcluster.config.yaml
@@ -23,7 +23,9 @@ Scheduling:
           {% else %}
           Instances:
             - InstanceType: {{ instance }}
-          # set MinCount to 1 in order to test compute node disk and memory metrics are not gathered
+          # set MinCount to 1 to ensure the existence of compute nodes
+          # in order to test compute node disk and memory metrics are not gathered
+          # and test detailed monitoring setting
           MinCount: 1
           {% endif %}
 Monitoring:
