+++ tests/integration-tests/tests/monitoring/test_structured_log_events/test_custom_compute_action_failure/on_compute_configured_error.sh
