+++ tests/integration-tests/tests/monitoring/test_structured_log_events/test_custom_compute_action_failure/pcluster.config.yaml
